# contracts/openapi/identity-service.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TradeDiary.Identity | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5600/"
    }
  ],
  "paths": {
...
    "/internal/auth/settings": {
      "get": {
        "tags": [
          "TradeDiary.Identity"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "TradeDiary.Identity"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/internal/auth/account-export": {
      "get": {
        "tags": [
          "TradeDiary.Identity"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountIdentityExport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/internal/auth/account": {
      "delete": {
        "tags": [
          "TradeDiary.Identity"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountAgentExport": {
        "required": [
          "userId",
          "displayName",
          "timezone",
          "baseCurrency",
          "createdAt",
          "keyId",
          "keyName",
          "scopes",
          "tokenCreatedAt",
          "revokedAt",
          "lastUsedAt",
          "lastSuccessfulRequestAt"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "baseCurrency": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "keyId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "keyName": {
            "type": [
              "null",
              "string"
            ]
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tokenCreatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "lastSuccessfulRequestAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "AccountDeletionRequest": {
        "required": [
          "confirmation"
        ],
        "type": "object",
        "properties": {
          "confirmation": {
            "type": "string"
          }
        }
      },
      "AccountIdentityExport": {
        "required": [
          "profile",
          "agents"
        ],
        "type": "object",
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/AccountProfileExport"
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AccountAgentExport"
            }
          }
        }
      },
      "AccountProfileExport": {
        "required": [
          "id",
          "email",
          "displayName",
          "timezone",
          "journalDayRollover",
          "baseCurrency",
          "appearance",
          "locale",
          "accentTheme",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "journalDayRollover": {
            "type": "string"
          },
          "baseCurrency": {
            "type": "string"
          },
          "appearance": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "accentTheme": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UserSettingsResponse": {
        "required": [
          "email",
          "displayName",
          "timezone",
          "journalDayRollover",
          "baseCurrency",
          "appearance",
          "locale",
          "accentTheme",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "journalDayRollover": {
            "type": "string"
          },
          "baseCurrency": {
            "type": "string"
          },
          "appearance": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "accentTheme": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSettingsWrite": {
        "required": [
          "displayName",
          "timezone",
          "journalDayRollover",
          "baseCurrency",
          "appearance",
          "locale",
          "accentTheme"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "journalDayRollover": {
            "type": "string"
          },
          "baseCurrency": {
            "type": "string"
          },
          "appearance": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "accentTheme": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "serviceKey": {
        "type": "apiKey",
        "name": "X-Service-Key",
        "in": "header"
      }
    }
  },
...
}